change int to longint in string plugin (obsolete)

+++ string/plugin/openmm-struna/platforms/opencl/src/OpenCLStrunaKernels.cpp
@@ -164,7 +164,7 @@ void OpenCLCalcStrunaForceKernel::beginComputation(bool includeForces, bool incl
 }
 
 void OpenCLCalcStrunaForceKernel::executeOnWorkerThread() {
-    int iteration = cl.getStepCount();
+    long int iteration = cl.getStepCount();
     double* rptr; // pointer to coordinate array
     int* aptr; // pointer to atom index array
     int i, j, ierr;
